RSSEC-263 refactor: rename token classe files, add readonly status to properties
