+++ src/server/token/anonymousTokenManager.ts
